+++ src/main/java/es/uned/grc/pfc/meteo/server/model/Station.java
@@ -30,8 +30,8 @@ public class Station extends AbstractVersionable <Integer> {
    private StationModel stationModel = null;
    private Set <Parameter> parameters = null;
    private Set <Variable> variables = null;
-   private Long latitude = null;
-   private Long longitude = null;
+   private Double latitude = null;
+   private Double longitude = null;
    private Integer height = null;
    private String zip = null;
    private String city = null;
@@ -85,18 +85,18 @@ public class Station extends AbstractVersionable <Integer> {
    }
    
    @Column
-   public Long getLatitude () {
+   public Double getLatitude () {
       return latitude;
    }
-   public void setLatitude (Long latitude) {
+   public void setLatitude (Double latitude) {
       this.latitude = latitude;
    }
    
    @Column
-   public Long getLongitude () {
+   public Double getLongitude () {
       return longitude;
    }
-   public void setLongitude (Long longitude) {
+   public void setLongitude (Double longitude) {
       this.longitude = longitude;
    }
